- Made Hologram work in 1.18.2 - UI still needs to be done for 1.18.2 - Copied http util from T++ - Code refactoring seriously needed

+++ btr-core/src/main/java/com/mndk/bteterrarenderer/connector/graphics/IBufferBuilder.java
@@ -3,7 +3,7 @@ package com.mndk.bteterrarenderer.connector.graphics;
 public interface IBufferBuilder {
     void beginQuads(VertexFormatConnectorEnum vertexFormat);
     IBufferBuilder pos(double x, double y, double z);
-    IBufferBuilder tex(double u, double v);
+    IBufferBuilder tex(float u, float v);
     IBufferBuilder color(float red, float green, float blue, float alpha);
     void endVertex();
 }
